api,params: change data size to size_t
Data type parameters are used in 2 node types: AnimKeyFrameBuffer and
all the Buffer* nodes.

AnimKeyFrameBuffer is using animkeyframe_opts to store the parameter,
which is shared amongst animation routines. AnimatedBuffer reads the
data field and transfers it to a buffer_info, so this struct needed an
update of its data_size as well.

Buffer* use a local buffer_opts, and similary to AnimatedBuffer
transfers the data field into a buffer_info.

+++ libnopegl/src/deserialize.c
@@ -335,18 +335,18 @@ static int parse_param_str(struct darray *nodes_array, uint8_t *dstp,
 static int parse_param_data(struct darray *nodes_array, uint8_t *dstp,
                             const struct node_param *par, const char *str)
 {
-    int size = 0;
+    size_t size = 0;
     int consumed = 0;
     const char *cur = str;
     const char *end = str + strlen(str);
-    int ret = sscanf(str, "%d,%n", &size, &consumed);
+    int ret = sscanf(str, "%zd,%n", &size, &consumed);
     if (ret != 1 || !size || cur >= end - consumed)
         return NGL_ERROR_INVALID_DATA;
     cur += consumed;
     uint8_t *data = ngli_calloc(size, sizeof(*data));
     if (!data)
         return NGL_ERROR_MEMORY;
-    for (int i = 0; i < size; i++) {
+    for (size_t i = 0; i < size; i++) {
         if (cur > end - 2) {
             ngli_free(data);
             return NGL_ERROR_INVALID_DATA;
